float type of numbers

+++ Assignment2/Assignment2b/head.h
@@ -1,6 +1,8 @@
 #ifndef HEAD_H
 #define HEAD_H
+#include <iostream>
+#include <string>
 int main();
 bool operand(std::string oper);
-void func(int *a, int *b, int *index_mass, std::string *z);
+void func(float *a, float *b, int *index_mass, std::string *z);
 #endif
